+++ DemoApp/DemoAppCoordinator.swift
@@ -141,7 +141,7 @@ private extension DemoAppCoordinator {
         let channelListVC = DemoChatChannelListVC.make(with: controller)
         channelListVC.demoRouter?.onLogout = onLogout
         channelListVC.selectedChannel = selectedChannel
-        channelListVC.components.isChatChannelListDefaultLoadingStatesEnabled = true
+        channelListVC.components.isChatChannelListStatesEnabled = true
         return channelListVC
     }