+++ Sources/StreamChatUI/Components.swift
@@ -312,9 +312,9 @@ public struct Components {
     /// View that shows when loading the Channel list.
     public var chatChannelListLoadingView: ChatChannelListLoadingView.Type = ChatChannelListLoadingView.self
     
-    /// A boolean value that determines whether the Channel list default loading states (empty, error and loading views) are shown in Channel list.
-    /// False by default not showing empty or error views and just showing a spinner indicator for loading state. If set true, default empty, error and shimmer loading views are shown instead.
-    public var isChatChannelListDefaultLoadingStatesEnabled = false
+    /// A boolean value that determines whether the Channel list default loading states (empty, error and loading views) are handled by the Stream SDK. It is false by default.
+    /// By default it is false, it does not show empty or error views and just shows a spinner indicator for the loading state. If set to true, the empty, error and shimmer loading views are shown instead.
+    public var isChatChannelListStatesEnabled = false
 
     // MARK: - Composer components