+++ Tests/StreamChatUITests/SnapshotTests/ChatChannelList/ChatChannelListVC_Tests.swift
@@ -60,7 +60,7 @@ final class ChatChannelListVC_Tests: XCTestCase {
     }
 
     func test_emptyAppearance() {
-        vc.components.isChatChannelListDefaultLoadingStatesEnabled = true
+        vc.components.isChatChannelListStatesEnabled = true
         vc.executeLifecycleMethods()
         mockedChannelListController.simulate(state: .remoteDataFetched)
         AssertSnapshot(vc, isEmbeddedInNavigationController: true)
@@ -90,7 +90,7 @@ final class ChatChannelListVC_Tests: XCTestCase {
         }
 
         var components = Components.mock
-        components.isChatChannelListDefaultLoadingStatesEnabled = false
+        components.isChatChannelListStatesEnabled = false
         components.channelCellSeparator = TestView.self
         vc.components = components
 
@@ -182,7 +182,7 @@ final class ChatChannelListVC_Tests: XCTestCase {
     func test_didChangeState_whenRemoteDataFetchedAndChannelsAreEmpty_thenEmptyViewIsShown() {
         // GIVEN
         let emptyViewHidden = false
-        vc.components.isChatChannelListDefaultLoadingStatesEnabled = true
+        vc.components.isChatChannelListStatesEnabled = true
         
         // WHEN
         vc.controller(vc.controller, didChangeState: .remoteDataFetched)
@@ -194,7 +194,7 @@ final class ChatChannelListVC_Tests: XCTestCase {
     func test_didChangeState_whenRemoteDataFetchedFailed_thenErrorViewIsShown() {
         // GIVEN
         let errorViewHidden = false
-        vc.components.isChatChannelListDefaultLoadingStatesEnabled = true
+        vc.components.isChatChannelListStatesEnabled = true
         
         // WHEN
         vc.controller(vc.controller, didChangeState: .remoteDataFetchFailed(ClientError()))